[FIXED] Incorrect number of subscriptions in /streaming/serverz
Replaces the use of a variable that was supposed to be updated
on subscription create/remove. It was started to be a bit brittle.

Provide a function that computes the number of subscriptions
(including offline durables/queue durables).
This is a bit more costly than using the variable but will be
less prone to issues.

Resolves #819

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>