[FIXED] Clustering: restore snapshot with some messages no longer avail
When a snapshot is performed prior to messages being removed (either
expired or remove due to count/size limits), a node that would
restore from this snapshot would ask the leader to send messages
which indexes (first/last) are in the snapshot.
Both leader and node would stop the restore when the first unavail
message was found.

Resolves #834

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>