Optimize restoreMsgsFromSnapshot and make it backward compatible
PR #835 fixed an issue with restoreMsgsFromSnapshot when there
were unavailable messages (expired or removed due to limits).
However, the implementation was not optimized when lots of
messages are no longer available. Also, it would not be backward
compatible with older servers.

This change detects if two servers are 0.14.1+ and can then
send the first available sequence if the leader does not find
the requested message.
With this code change, a follower that requests messages to a
leader that is of previous version will handle unavailable
messages in an optimized way. It will first send few requests
starting at the next sequence but then will try to do a binary
search but first ask the leader for the last message on its
range. If this message is not found, then that means all messages
in the range have expired.

After several releases, when we deem reasonable that all servers
should be at 0.14.2+, we could remove some of the complexity
in this PR.

Resolves #834

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>