[FIXED] Redelivery of unack'ed msgs of leaving queue member
When a queue member has unacknowledged messages and is leaving the
group, these messages are supposed to be immediately redelivered
to existing queue members. However, if a target member had unack'ed
messages of a sequence lower than the reassigned messages these
reassigned messages would not be redelivered until after the AckWait.

The issue was introduced in PR https://github.com/nats-io/nats-streaming-server/pull/729
with the incorrect change in ordering algo for the messages that
should expire.

Resolves #869

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>