[ADDED] FileStore: read ahead capability
A new option `file_read_buffer_size` is added and defaults to 2MB.
If not set to 0, when a message is read from disk, the store will
actually read the following messages, up to that buffer size, and
add them to the internal cache (which already existed and has a
default TTL of 1sec).
This will speed up lookups of following messages since they are
likely to be in the cache.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>