Fixed implementation of FileStore auto-sync feature
The original implementation was having a central timer routine
that would go through the list of channels (interface) and cast
them to FileMsgStore and FileSubStore to invoke the stores' autoSync()
functions.
However, with stores that wrap a FileStore instance (such as RAFT or
Crypto store), that would cause a panic since the casting would fail.

Alternative fix would have been for FileStore to maintain a separate
map of actual File[Msg/Sub]Store, but I chose instead to move the
call to autoSync() into the Msg/Sub stores themselves.

Relates to #882

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>