[FIXED] Ensure Subscription response sent prior to messages
Original issue was addressed in PR #247, but unfortunately the fix
did not take into account durable subscriptions. If a durable
subscription is restarted, the subscription object is in memory
and would have the flag "initialized" already set to true, which
would not prevent published messages at the time of the subscription
request processing to allow messages to be sent prior to the
subscription response.

Resolves #930

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>