[FIXED] Clustering: possible wrong pending_count on followers
This happens when a queue subscriber with pending message is
being removed and messages are reassigned to other members. Each
node was doing this operation resulting possibly in messages
being "assigned" to different qsubs.
After a new leader election, it would be possible for messages
to be redelivered, due to this.

Resolves #934

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>