[FIXED] Cluster: Handling of deleted channels
A channel could be deleted and then later recreated (same name).
This is not handled properly in some cases such as restore from
snapshot.
This PR attempts to fix the problem by storing in the RAFT db
a ID when a channel is created. This ID is added to some of
the replicated operations.

Resolves #1008
Resolves #1010

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>