[FIXED] Windows Service: do not force usage of syslog
When starting as a Windows Service, we were creating a syslog to
possibly log the failure to start. But users running without
admin privileges would be unable to use the service since the
server would fail to create the event log entry.
So create the syslog only if explicit set (-syslog <bool> option)
or if no logfile has been specified.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>