[ADDED] Reload signal now accepted for the embedded NATS Server
Previously, the reload signal was ignored. It will now be passed
on to the embedded NATS Server. If the streaming server remotely
connects to a NATS Server and a reload signal is received, the
streaming server will print a warning indicating that it works
only for NATS Server configuration and only if it is embedded.

Resolves #1030

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>