[FIXED] FileStore: index file may be recreated with wrong offsets
This could happen if recovery of an index does not have matching
information with the last message in corresponding ".dat" file.
In that case, the server wipes the index file and recreate it
based on the data file. However, the offset was not reset, which
means that the index file would contain wrong offsets.

Resolves #1086

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>