[FIXED] Possible race causing ack'ed message to be redelivered
During redelivery, a go routine gathers all pending messages under
a lock. However, the actual send of messages is done one by one
and by then, one of the message in the list of messages to redeliver
may have been acknowledged. So during the send, under the sub's lock,
we need to check that if a message marked as redelivered is no longer
in the pending list, it should not be sent to the client.

Also fixed an unrelated data race that was just found running tests.

Resolves #1138

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>