[FIXED] Clusering: don't report "sub close request missing subject" on replay
Suppose that a channel exists and there is a subscription for this
channel. A snapshot occurs, then later the subscription is closed.
After a while, the channel is removed.

If a server restarts and replays the snapshot and the rest of the
RAFT log, it will skip the recovery of the channel because it
would detect that the channel no longer exists, however, the
replay of the log would try to process the subscription close
and not find the channel.
We detect this case now and simply ignore the sub close protocol.

We also changed the error message, that still can be reported when
processing the request in the leader (before replication), from
"missing subject" to "unknown channel".

Resolves #1229

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>