[FIXED] Recover channels even if server information not recovered
A user reported that the server.dat was missing, which means that
the server did not recover channels (assuming that this was a
fresh start, without any state). This caused issues later on
since if an app publishes a message to an existing (non recovered)
channel, then messages would be added to the channel but with
a reset of the sequence.

The server will now warn of the situation, initialize the server
state and then recover the rest of the state (channels, etc..)

Resolves #1263

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>